Handle the alternate data format used by local mqtt
The latest glow cad firmware supports local mqtt. Unfortunately
the data format used by local mqtt is very different than the
hosted mqtt. This commit adds support for the local mqtt format
in addition to the hosted format.

When running locally, username and password should be set to
empty strings if not needed.
If TLS is not used, the `ca_certs` should be set to empty string.

The data format published to influx is slightly different:
- `gid` label will be None, since this is no longer published
  in the sensor readings message
- unit rate and standing charge metrics will be published for
  both gas and electricity
- `gas_meter_reading_kwh` will be published, which contains
  the pre-converted meter reading.
- All metrics are passed through as-is from the mqtt data
  without processing values